--- queue-manager-basic-deployment/code/pom.xml
@@ -1,15 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.ibm.integration</groupId>
-  <artifactId>jms</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>producer</name>
-    <!-- https://mvnrepository.com/artifact/com.ibm.mq/com.ibm.mq.allclient -->
-<dependencies>
-<dependency>
-    <groupId>com.ibm.mq</groupId>
-    <artifactId>com.ibm.mq.allclient</artifactId>
-    <version>9.3.3.0</version>
-</dependency>
-</dependencies>
-</project>
