+++ queue-manager-basic-deployment/code/qmdemo-producer/src/main/java/com/ibm/integration/qmdemo/App.java
@@ -1,0 +1,13 @@
+package com.ibm.integration.qmdemo;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
