+++ queue-manager-basic-deployment/code/qmdemo-producer/src/test/java/com/ibm/integration/AppTest.java
@@ -1,0 +1,20 @@
+package com.ibm.integration;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
